refactor(csv): move csv parsing into separate file
