+++ lib/yargs/commands/file.js
@@ -20,7 +20,7 @@ module.exports = () => {
           const eventQRCode = createEventQRCode({ locationData, vendorData })
           const ext = path.extname(filepath).substr(1)
           if (!supportedExtensions.includes(ext)) {
-            throw new Error(`File extension ${ext} not in ${supportedExtensions}`)
+            throw new Error(`File extension ${ext} of ${filepath} not in ${supportedExtensions}`)
           }
 
           const absFilepath = path.resolve(process.cwd(), filepath)
